test: extract shared GitCli helper for TestTriggerMap fixtures
The new TestTriggerMap test files each hand-rolled the same
ProcessStartInfo("git") plumbing to run git for arrange-phase fixture
setup: a near-identical RunGit in SelectTestsCliTests and
GraphAffectedProjectsTests, plus an inline git ls-files in
TestTriggerMapTests.

Consolidate into a small synchronous Infrastructure.Tests/Shared/GitCli
helper (Run(workingDirectory, params args) -> trimmed stdout, throws with
stderr on non-zero exit). The three callers now delegate to it; the
per-file RunGit/Git wrappers stay as thin working-directory-bound
conveniences so call sites are unchanged.

Deliberately a small blocking helper rather than the async
NodeCommand/PowerShellCommand abstractions: those run the script under
test with output streaming, timeouts, and diagnostics, whereas git here
is just scaffolding.
