+++ tests/Infrastructure.Tests/TestTriggerMap/SelectTestsCliTests.cs
@@ -1,7 +1,6 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-using System.Diagnostics;
 using System.Text.Json;
 using Xunit;
 
@@ -790,30 +789,5 @@ public sealed class SelectTestsCliTests
         RunGit(repoRoot, "commit", "-q", "-m", message);
     }
 
-    private static string RunGit(string repoRoot, params string[] args)
-    {
-        var psi = new ProcessStartInfo("git")
-        {
-            WorkingDirectory = repoRoot,
-            RedirectStandardOutput = true,
-            RedirectStandardError = true,
-            UseShellExecute = false,
-        };
-        psi.ArgumentList.Add("--no-pager");
-        foreach (var arg in args)
-        {
-            psi.ArgumentList.Add(arg);
-        }
-
-        using var process = Process.Start(psi) ?? throw new InvalidOperationException("Failed to start git.");
-        var stdout = process.StandardOutput.ReadToEnd();
-        var stderr = process.StandardError.ReadToEnd();
-        process.WaitForExit();
-        if (process.ExitCode != 0)
-        {
-            throw new InvalidOperationException($"git {string.Join(' ', args)} failed ({process.ExitCode}): {stderr}");
-        }
-
-        return stdout.Trim();
-    }
+    private static string RunGit(string repoRoot, params string[] args) => GitCli.Run(repoRoot, args);
 }
